Reincarnated main_test.go but just with the TestMain() func.
